@@ -1,9 +1,10 @@
 buildscript {
     repositories {
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.0-beta6'
+        classpath 'com.android.tools.build:gradle:3.2.0-rc02'
         classpath 'com.novoda:bintray-release:0.4.0'
     }
 }
@@ -11,13 +12,14 @@ buildscript {
 allprojects {
     repositories {
         jcenter()
+        google()
     }
 }
 
 ext {
-    compileSdkVersion = 25
-    buildToolsVersion = "26.0.1"
-    supportLibVersion = "25.3.3"
+    compileSdkVersion = 28
+    appcompatVersion = "1.0.0-rc02"
+    cardviewVersion = "1.0.0-rc02"
 }
 
 task clean(type: Delete) {
